Increase coverage to 100%
