+++ tests/test_init.py
@@ -1,7 +1,5 @@
-"""Test pysma sensors."""
-# from homeassistant.loader import async_get_dhcp
+"""Test pysma init."""
 import logging
-from json import loads
 from unittest.mock import patch
 
 import aiohttp
@@ -11,135 +9,20 @@ from pysma import SMA
 from pysma.const import (
     DEVCLASS_INVERTER,
     ENERGY_METER_VIA_INVERTER,
-    JMESPATH_VAL,
-    JMESPATH_VAL_IDX,
     OPTIMIZERS_VIA_INVERTER,
 )
 from pysma.definitions import sensor_map
+from pysma.definitions import status as status_sensor
 from pysma.exceptions import (
     SmaAuthenticationException,
     SmaConnectionException,
     SmaReadException,
 )
-from pysma.sensor import Sensor, Sensors
+from pysma.sensor import Sensors
 
 from . import MOCK_DEVICE, MOCK_L10N, mock_aioresponse  # noqa: F401
 
 _LOGGER = logging.getLogger(__name__)
-
-
-SB_1_5 = loads(
-    # {"result": {"012F-7309879F":
-    """
-    {
-        "6400_00260100": {"1": [{"val": 3514000}]},
-        "6400_00262200": {"1": [{"val": 402}]}
-    }
-    """
-    # }}
-)
-
-SB_2_5 = loads(
-    # {"result": {"012F-730A2A43":
-    """
-    {
-        "6400_00262200": {"val": 402},
-        "6100_40263F00": {"val": null},
-        "6400_00260100": {"val": 3514000}
-    }
-    """
-    # }}
-)
-
-
-@pytest.fixture
-def sensors():
-    """Fixture to create some sensors."""
-    yield [
-        (402, True, Sensor("6400_00262200", "s_402", "W")),
-        (3514, True, Sensor("6400_00260100", "s_3514", "W", 1000)),
-    ]
-
-
-class Test_sensor_class:
-    """Test the Sensor class."""
-
-    def test_sensor_sb_1_5(self, sensors):
-        """Test extract value."""
-        for val, change, sens in sensors:
-            assert sens.path is None
-            assert sens.extract_value(SB_1_5) is change
-            assert sens.value == val
-            assert sens.path == JMESPATH_VAL_IDX.format("1", sens.key_idx)
-
-            assert sens.extract_value(SB_1_5) is False
-
-    def test_sensor_sb_2_5(self, sensors):
-        """Test extract value."""
-        for val, change, sens in sensors:
-            assert sens.path is None
-            assert sens.extract_value(SB_2_5) is change
-            assert sens.value == val
-            assert sens.path == JMESPATH_VAL
-
-            assert sens.extract_value(SB_2_5) is False
-
-    def test_null(self):
-        """Test a null or None result."""
-        sens = Sensor("6100_40263F00", "s_null", "W")
-        assert (
-            sens.extract_value({"result": {"_": {"6100_40263F00": {"val": None}}}})
-            is False
-        )
-        assert sens.value is None
-        assert (
-            sens.extract_value(
-                {"result": {"_": {"6100_40263F00": {"1": [{"val": None}]}}}}
-            )
-            is False
-        )
-        assert sens.value is None
-        assert sens.extract_value({"result": {"_": {}}}) is False
-        assert sens.value is None
-
-    @patch("pysma.sensor._LOGGER.warning")
-    def test_default_no_duplicates(self, mock_warn):
-        """Ensure warning on duplicates."""
-        sen = Sensors(sensor_map[DEVCLASS_INVERTER])
-        assert len(sen) > 20
-        assert len(sen) < 50
-        assert mock_warn.call_count == 0
-        # Add duplicate frequency
-        news = Sensor("key1", "frequency", "")
-        sen.add(news)
-        assert mock_warn.call_count == 1
-        assert sen[news.name] == news
-        # Add duplicate freq, key should not be raised
-        sen.add(Sensor("6100_00465700", "frequency", ""))
-        assert mock_warn.call_count == 2
-        # Add duplicate freq key only
-        sen.add(Sensor("6100_00465700", "f001", ""))
-        assert mock_warn.call_count == 3
-        # Test different key_idx only
-        sen.add(Sensor("key1_0", "frequency_0", ""))
-        assert mock_warn.call_count == 3
-        sen.add(Sensor("key1_1", "frequency_1", ""))
-        assert mock_warn.call_count == 3
-
-    @patch("pysma.sensor._LOGGER.warning")
-    def test_type_error(self, mock_warn):
-        """Ensure TypeError on not isinstance."""
-        sen = Sensors()
-        with pytest.raises(TypeError):
-            sen.add("This is not a Sensor")
-
-    @patch("pysma.sensor._LOGGER.warning")
-    def test_default_jmes(self, mock_warn):
-        """Ensure default sensors are ok."""
-        sens = Sensors(sensor_map[DEVCLASS_INVERTER])
-        for sen in sens:
-            sen.extract_value(SB_1_5)
-        assert mock_warn.called
 
 
 class Test_SMA_class:
@@ -182,6 +65,31 @@ class Test_SMA_class:
         assert isinstance(json, dict)
         assert json == {}
         assert mock_warn.call_count == 1
+
+    @patch("pysma._LOGGER.warning")
+    async def test_read_no_password(self, mock_warn, mock_aioresponse):  # noqa: F811
+        """Test read_body without password."""
+        mock_aioresponse.post(
+            f"{self.base_url}/dyn/getDashValues.json",
+            payload={
+                "result": {
+                    "0199-xxxxx385": {
+                        "6800_08822000": {
+                            "1": [
+                                {
+                                    "validVals": [9401, 9402, 9403, 9404, 9405],
+                                    "val": [{"tag": 9402}],
+                                }
+                            ]
+                        },
+                    }
+                }
+            },
+        )
+        session = aiohttp.ClientSession()
+        sma = SMA(session, self.host)
+        await sma.read(Sensors(status_sensor))
+        assert mock_warn.call_count == 0
 
     @patch("pysma._LOGGER.warning")
     async def test_read_body_error(self, mock_warn, mock_aioresponse):  # noqa: F811
@@ -335,10 +243,16 @@ class Test_SMA_class:
 
     async def test_new_session_error(self, mock_aioresponse):  # noqa: F811
         """Test new_session with error."""
-        mock_aioresponse.post(f"{self.base_url}/dyn/login.json", payload={"err": {}})
-
         session = aiohttp.ClientSession()
         sma = SMA(session, self.host, "pass")
+
+        mock_aioresponse.post(
+            f"{self.base_url}/dyn/login.json", payload={"err": "dummy-error"}
+        )
+        with pytest.raises(SmaAuthenticationException):
+            await sma.new_session()
+
+        mock_aioresponse.post(f"{self.base_url}/dyn/login.json", payload={"err": 503})
         with pytest.raises(SmaAuthenticationException):
             await sma.new_session()
